Made users unable to edit other user's infos and comments, also made sign in a requirement to view pages

+++ app/controllers/comments_controller.rb
@@ -4,6 +4,9 @@ class CommentsController < ApplicationController
   end
 
   def edit
+    unless @current_account.id == params[:id]
+      redirect_to comments_path
+    end
     @comment = Comment.find params[:id]
   end
 
@@ -32,6 +35,9 @@ class CommentsController < ApplicationController
   end
 
   def destroy
+    unless @current_account.id == params[:id]
+      redirect_to comments_path
+    end
     comment = Comment.find params[:id]
     comment.destroy
     redirect_to comments_path
